requirements.txt
langchain>=0.1.0
langchain-google-genai>=0.0.11
google-generativeai>=0.3.0
autopep8>=2.0.0
python-dotenv>=1.0.0
datasets>=2.14.0